pyproject.toml
[project]
name = "voiceagent"
version = "0.1.0"
description = "Cosentus voice agent platform — Pipecat + Twilio + Supabase"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.0",
    "pipecat-ai-small-webrtc-prebuilt>=2.4.0",
    "pipecat-ai-whisker>=0.0.11",
    "pipecat-ai[anthropic,cartesia,deepgram,elevenlabs,livekit,local,silero,turn-detector,webrtc,websocket]>=0.0.102",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "supabase>=2.28.3",
    "twilio>=9.10.3",
    "uvicorn[standard]>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pyyaml>=6.0.3",
]
